+++ examples/hn.svelte.dev/.netlify/netlify-plugin-pnpm/index.js
@@ -1,8 +1,9 @@
-export default {
+module.exports = {
   onPreBuild: async ({ utils: { build, run } }) => {
     try {
       await run.command("npm install -g pnpm")
-      await run.command("pnpm install")
+      await run.command("pnpm -w install")
+      await run.command("pnpm -w build")
     } catch (error) {
       return build.failBuild(error)
     }