+++ packages/adapter-node/rollup.config.js
@@ -4,6 +4,16 @@ import json from '@rollup/plugin-json';
 
 export default [
 	{
+		input: 'src/middlewares.js',
+		output: {
+			file: 'files/middlewares.js',
+			format: 'esm',
+			sourcemap: true
+		},
+		plugins: [nodeResolve(), commonjs(), json()],
+		external: ['../output/server/app.js', ...require('module').builtinModules]
+	},
+	{
 		input: 'src/index.js',
 		output: {
 			file: 'files/index.js',
@@ -11,7 +21,7 @@ export default [
 			sourcemap: true
 		},
 		plugins: [nodeResolve(), commonjs(), json()],
-		external: ['../output/server/app.js', './env.js', ...require('module').builtinModules]
+		external: ['./middlewares.js', './env.js', ...require('module').builtinModules]
 	},
 	{
 		input: 'src/shims.js',