+++ packages/adapter-node/tests/smoke.js
@@ -1,14 +1,15 @@
 import { test } from 'uvu';
-import { createServer } from '../src/server.js';
+import { create_kit_middleware } from '../src/kit-middleware.js';
 import * as assert from 'uvu/assert';
 import fetch from 'node-fetch';
+import polka from 'polka';
 
 const { PORT = 3000 } = process.env;
 const DEFAULT_SERVER_OPTS = { render: () => {} };
 
-function startServer(opts = DEFAULT_SERVER_OPTS) {
-	const server = createServer(opts);
+async function startServer(opts = DEFAULT_SERVER_OPTS) {
 	return new Promise((fulfil, reject) => {
+		const server = polka().use(create_kit_middleware(opts));
 		server.listen(PORT, (err) => {
 			if (err) {
 				reject(err);