+++ packages/kit/src/core/config/index.spec.js
@@ -87,6 +87,15 @@ test('errors on invalid nested values', () => {
 	}, /^Unexpected option config\.kit\.files\.potato$/);
 });
 
+test('does not error on invalid top-level values', () => {
+	assert.not.throws(() => {
+		validate_config({
+			// @ts-expect-error - valid option for others but not in our definition
+			onwarn: () => {}
+		});
+	});
+});
+
 test('errors on extension without leading .', () => {
 	assert.throws(() => {
 		validate_config({