+++ packages/kit/src/packaging/test/fixtures/javascript_no_types/svelte.config.cjs
@@ -1,7 +1,0 @@
-module.exports = {
-	kit: {
-		package: {
-			emitTypes: false
-		}
-	}
-};